requirements.txt
flask
flask-cors
docxtpl
Pillow
python-docx
